requirements.txt
asyncpg
aiogram==2.10.1
aiohttp>=3.6.2
aioredis>=1.3.1
python-dotenv